+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/index.ts
@@ -32,8 +32,8 @@ export * from './phone-uk';
 export * from './query-management';
 export * from './text';
 export * from './text-area';
+export * from './rich-text-area';
 export * from './unsupported-field.component';
 export * from './utils';
 export * from './waystopay';
 export * from './yes-no';
-